Compiling after merge from v2-spec

+++ rust/pact_consumer/src/mock_server.rs
@@ -31,7 +31,7 @@ pub struct ValidatingMockServer {
     // A description of our mock server, for use in error messages.
     description: String,
     // The ID of our mock server.
-    port: i32,
+    port: u16,
     // The URL of our mock server.
     url: Url,
 }
